chatbot/pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.3",
    "flask-cors>=6.0.2",
    "langchain>=1.2.17",
    "langchain-core>=1.3.2",
    "langchain-openai>=1.2.1",
    "python-dotenv>=1.2.2",
]
